[update] todoContent test
